tree-like structure for tests that mirrors the library module hierarchy
